+++ pyfakefs/fake_filesystem_unittest.py
@@ -85,7 +85,7 @@ else:
 
 def load_doctests(loader, tests, ignore, module,
                   additional_skip_names=None,
-                  patch_path=True, special_names=None):  # pylint: disable=unused-argument
+                  patch_path=True):  # pylint: disable=unused-argument
     """Load the doctest tests for the specified module into unittest.
         Args:
             loader, tests, ignore : arguments passed in from `load_tests()`
@@ -96,7 +96,7 @@ def load_doctests(loader, tests, ignore, module,
     File `example_test.py` in the pyfakefs release provides a usage example.
     """
     _patcher = Patcher(additional_skip_names=additional_skip_names,
-                       patch_path=patch_path, special_names=special_names)
+                       patch_path=patch_path)
     globs = _patcher.replace_globs(vars(module))
     tests.addTests(doctest.DocTestSuite(module,
                                         globs=globs,
@@ -111,7 +111,7 @@ class TestCase(unittest.TestCase):
     """
 
     def __init__(self, methodName='runTest', additional_skip_names=None,
-                 patch_path=True, special_names=None,
+                 patch_path=True,
                  modules_to_reload=None,
                  use_dynamic_patch=True):
         """Creates the test class instance and the stubber used to stub out
@@ -129,11 +129,6 @@ class TestCase(unittest.TestCase):
 
                 Irrespective of patch_path, module 'os.path' is still correctly faked
                 if imported the usual way using `import os` or `import os.path`.
-            special_names: A dictionary with module names as key and a dictionary as
-                value, where the key is the original name of the module to be patched,
-                and the value is the name as it is imported.
-                This allows to patch modules where some of the file system modules are
-                imported as another name (e.g. `import os as _os`).
             modules_to_reload (experimental): A list of modules that need to be reloaded
                 to be patched dynamically; may be needed if the module
                 imports file system modules under an alias
@@ -155,20 +150,13 @@ class TestCase(unittest.TestCase):
                 def __init__(self, methodName='runTest'):
                     super(MyTestCase, self).__init__(
                         methodName=methodName, additional_skip_names=['posixpath'])
-
-
-            class AnotherTestCase(fake_filesystem_unittest.TestCase):
-                def __init__(self, methodName='runTest'):
-                    # allow patching a module that imports `os` as `my_os`
-                    special_names = {'amodule': {'os': 'my_os'}}
-                    super(MyTestCase, self).__init__(
-                        methodName=methodName, special_names=special_names)
         """
         super(TestCase, self).__init__(methodName)
         self._stubber = Patcher(additional_skip_names=additional_skip_names,
-                                patch_path=patch_path,
-                                special_names=special_names)
-        self._modules_to_reload = modules_to_reload or []
+                                patch_path=patch_path)
+        self._modules_to_reload = [tempfile]
+        if modules_to_reload is not None:
+            self._modules_to_reload.extend(modules_to_reload)
         self._use_dynamic_patch = use_dynamic_patch
 
     @property
@@ -223,15 +211,19 @@ class TestCase(unittest.TestCase):
         """
         self._stubber.setUp()
         self.addCleanup(self._stubber.tearDown)
+        dyn_patcher = DynamicPatcher(self._stubber)
+        sys.meta_path.insert(0, dyn_patcher)
 
         for module in self._modules_to_reload:
             if module.__name__ in sys.modules:
                 reload(module)
+
         if self._use_dynamic_patch:
-            dyn_patcher = DynamicPatcher(self._stubber)
-            sys.meta_path.insert(0, dyn_patcher)
             self.addCleanup(lambda: sys.meta_path.pop(0))
             self.addCleanup(dyn_patcher.cleanup)
+        else:
+            dyn_patcher.cleanup()
+            sys.meta_path.pop(0)
 
     @DeprecationWarning
     def tearDownPyfakefs(self):
@@ -268,13 +260,10 @@ class Patcher(object):
     if HAS_PATHLIB:
         SKIPNAMES.add('pathlib')
 
-    def __init__(self, additional_skip_names=None, patch_path=True,
-                 special_names=None):
+    def __init__(self, additional_skip_names=None, patch_path=True):
         """For a description of the arguments, see TestCase.__init__"""
 
         self._skipNames = self.SKIPNAMES.copy()
-        self._special_names = special_names or {}
-        self._special_names['tempfile'] = {'os': '_os', 'io': '_io'}
 
         if additional_skip_names is not None:
             self._skipNames.update(additional_skip_names)
@@ -343,12 +332,6 @@ class Patcher(object):
             for name in self._modules:
                 if inspect.ismodule(module.__dict__.get(name)):
                     self._modules[name].add((module, name))
-            if '__name__' in module.__dict__ and module.__name__ in self._special_names:
-                module_names = self._special_names[module.__name__]
-                for name in self._modules:
-                    if name in module_names:
-                        if inspect.ismodule(module.__dict__.get(module_names[name])):
-                            self._modules[name].add((module, module_names[name]))
 
     def _refresh(self):
         """Renew the fake file system and set the _isStale flag to `False`."""
@@ -362,28 +345,7 @@ class Patcher(object):
         self._fake_modules['path'] = self._fake_modules['os'].path
         self.fake_open = fake_filesystem.FakeFileOpen(self.fs)
 
-        if not self.IS_WINDOWS and 'tempfile' in sys.modules:
-            self._patch_tempfile()
-
         self._isStale = False
-
-    def _patch_tempfile(self):
-        """Hack to work around cached `os` functions in `tempfile`.
-         Shall be replaced by a more generic mechanism.
-        """
-        if 'unlink' in tempfile._TemporaryFileWrapper.__dict__:
-            # Python 2.7 to 3.2: unlink is a class method of _TemporaryFileWrapper
-            tempfile._TemporaryFileWrapper.unlink = self._fake_modules['os'].unlink
-
-            #  Python 3.0 to 3.2 (and PyPy3 based on Python 3.2):
-            # `TemporaryDirectory._rmtree` is used instead of `shutil.rmtree`
-            # which uses several cached os functions - replace it with `shutil.rmtree`
-            if 'TemporaryDirectory' in tempfile.__dict__:
-                tempfile.TemporaryDirectory._rmtree = lambda o, path: shutil.rmtree(path)
-        else:
-            # Python > 3.2 - unlink is a default parameter of _TemporaryFileCloser
-            tempfile._TemporaryFileCloser.close.__defaults__ = (
-            self._fake_modules['os'].unlink,)
 
     def setUp(self, doctester=None):
         """Bind the file-related modules to the :py:mod:`pyfakefs` fake