+++ CHANGES.md
@@ -8,6 +8,7 @@ must use pyfakefs 3.3 or earlier.
 
 #### New Features
   * Added support to fake out backported `scandir` module ([#332](../../issues/332))
+  * Dynamic loading of modules after setup is now on by default and no more considered experimental (see [#340](../../issues/340)).
 
 #### Infrastructure
   * Changed API to be PEP-8 conform [#186](../../issues/186). Note: The old API is still available.